Fixed isnan for windows and added DDS_ALIVE as an attribute

+++ plotjuggler_plugins/ParserROS/rosx_introspection/include/rosx_introspection/details/conversion_impl.hpp
@@ -235,7 +235,7 @@ template<typename SRC,typename DST,
           typename details::EnableIf< details::unsigned_to_smaller_conversion<SRC, DST>>* = nullptr  >
 inline void convert_impl( const SRC& from, DST& target )
 {
-  if(!std::isnan(from)) {
+  if(!std::isnan(static_cast<double>(from))) {
     checkUpperLimit<SRC,DST>(from);
   }
   target = static_cast<DST>(from);
@@ -245,7 +245,7 @@ template<typename SRC,typename DST,
           typename details::EnableIf< details::signed_to_smaller_conversion<SRC, DST>>* = nullptr  >
 inline void convert_impl( const SRC& from, DST& target )
 {
-  if(!std::isnan(from)) {
+  if(!std::isnan(static_cast<double>(from))) {
     checkLowerLimit<SRC,DST>(from);
     checkUpperLimit<SRC,DST>(from);
   }
@@ -257,7 +257,7 @@ template<typename SRC,typename DST,
           typename details::EnableIf< details::signed_to_smaller_unsigned_conversion<SRC, DST>>* = nullptr  >
 inline void convert_impl( const SRC& from, DST& target )
 {
-  if (!std::isnan(from) && from < 0 ){
+  if (!std::isnan(static_cast<double>(from)) && from < 0 ){
     throw RangeException("Value is negative and can't be converted to signed");
   }
 
@@ -332,7 +332,7 @@ template<typename SRC,typename DST,
           typename details::EnableIf< details::integer_to_floating_conversion<SRC, DST>>* = nullptr >
 inline void convert_impl( const SRC& from, DST& target )
 {
-  if(!std::isnan(from)) {
+  if(!std::isnan(static_cast<double>(from))) {
     checkTruncation<SRC,DST>(from);
   }
   target = static_cast<DST>( from);
